relocated enums
