+++ ClassLibraryForHT9/Models/Product.cs
@@ -1,4 +1,4 @@
-﻿namespace ClassLibraryForHT9
+﻿namespace ClassLibraryForHT9.Models
 {
     public class Product
     {
