+++ ProductInventoryProjectUsingClasses/Program.cs
@@ -1,7 +1,8 @@
-﻿using ProductInventoryProjectUsingClasses.Models;
-using static testRepo.Programm;
-using Product = ClassLibraryForHT9.Product;
-using ProductInventory = ClassLibraryForHT9.ProductInventory;
+﻿using static testRepo.Programm;
+using Product = ClassLibraryForHT9.Models.Product;
+using ProductInventory = ClassLibraryForHT9.Models.ProductInventory;
+using MainMenuEnum = ClassLibraryForHT9.Models.Enums.MainMenuEnum;
+using ProductCreationMenuEnum = ClassLibraryForHT9.Models.Enums.ProductCreationMenuEnum;
 
 List<Product> allProducts = new()
 {
